Frida registration (Sign in)
Change the controllers as needed, define the controllers to used and
work with the layout view to hold the necessary fields.

NOT WORKING AS EXPECTED THO. It does create a new Team at the moment of
the Frida creation but it does not assign that Team to the Frida.
Also, it only performs the registration without considering whether she
does not have a team, wants to create a new team or has a team already.